Чистка контроллеров
Вынес паралакс в слой анимации в отдельный класс. Исправил в Main имя переменной с цветом. Убрал UserDefaults для цвета. Удалил закоментированный код с анимацией
